uniswap correction

+++ src/L2/Mock/Tokamak-uniswap/uniswap-core/libraries/FullMath.sol
@@ -12,17 +12,17 @@ library FullMath {
     /// @return result The 256-bit result
     /// @dev Credit to Remco Bloemen under MIT license https://xn--2-umb.com/21/muldiv
     function mulDiv(
-        uint256 a,
-        uint256 b,
-        uint256 denominator
-    ) internal pure returns (uint256 result) {
+        int256 a,
+        int256 b,
+        int256 denominator
+    ) internal pure returns (int256 result) {
         // 512-bit multiply [prod1 prod0] = a * b
         // Compute the product mod 2**256 and mod 2**256 - 1
         // then use the Chinese Remainder Theorem to reconstruct
         // the 512 bit result. The result is stored in two 256
         // variables such that product = prod1 * 2**256 + prod0
-        uint256 prod0; // Least significant 256 bits of the product
-        uint256 prod1; // Most significant 256 bits of the product
+        int256 prod0; // Least significant 256 bits of the product
+        int256 prod1; // Most significant 256 bits of the product
         assembly {
             let mm := mulmod(a, b, not(0))
             prod0 := mul(a, b)
@@ -61,7 +61,7 @@ library FullMath {
         // Factor powers of two out of denominator
         // Compute largest power of two divisor of denominator.
         // Always >= 1.
-        uint256 twos = denominator & (~denominator + 1);
+        int256 twos = denominator & (~denominator + 1);
         // Divide denominator by power of two
         assembly {
             denominator := div(denominator, twos)
@@ -84,7 +84,7 @@ library FullMath {
         // modulo 2**256 such that denominator * inv = 1 mod 2**256.
         // Compute the inverse by starting with a seed that is correct
         // correct for four bits. That is, denominator * inv = 1 mod 2**4
-        uint256 inv = (3 * denominator) ^ 2;
+        int256 inv = (3 * denominator) ^ 2;
         // Now use Newton-Raphson iteration to improve the precision.
         // Thanks to Hensel's lifting lemma, this also works in modular
         // arithmetic, doubling the correct bits in each step.
@@ -111,13 +111,13 @@ library FullMath {
     /// @param denominator The divisor
     /// @return result The 256-bit result
     function mulDivRoundingUp(
-        uint256 a,
-        uint256 b,
-        uint256 denominator
-    ) internal pure returns (uint256 result) {
+        int256 a,
+        int256 b,
+        int256 denominator
+    ) internal pure returns (int256 result) {
         result = mulDiv(a, b, denominator);
-        if (mulmod(a, b, denominator) > 0) {
-            require(result < type(uint256).max);
+        if (mulmod(uint256(a), uint256(b), uint256(denominator)) > 0) {
+            require(result < type(int256).max);
             result++;
         }
     }
